+++ source/artifacts/crystalshard.js
@@ -1,7 +1,7 @@
 const { ArtifactTemplate } = require("../classes");
 
 module.exports = new ArtifactTemplate("Crystal Shard",
-	"Spells have a @{0.85^copies*-1+1*100}% chance to avoid using durability when used.",
-	"Increase chance of saving durability (multiplicatively) by 15% per shard",
+	"Increase the blast range of Spells by @{copies}. Single target Spells become Blast @{copies}.",
+	"Increase range boost by 1 per shard",
 	"Untyped"
 );