+++ source/gear/prismaticblast-base.js
@@ -1,0 +1,27 @@
+const { GearTemplate } = require('../classes');
+const { SAFE_DELIMITER } = require('../constants');
+const { dealDamage } = require('../util/combatantUtil');
+
+module.exports = new GearTemplate("Prismatic Blast",
+	"Strike a foe and adjacent foes for @{damage} @{element} damage",
+	"Damage x@{critMultiplier}",
+	"Spell",
+	"Light",
+	200,
+	(targets, user, isCrit, adventure) => {
+		const { element, damage, critMultiplier } = module.exports;
+		let pendingDamage = user.getPower() + damage;
+		if (user.element === element) {
+			targets.forEach(target => {
+				target.addStagger("elementMatchFoe");
+			})
+		}
+		if (isCrit) {
+			pendingDamage *= critMultiplier;
+		}
+		return dealDamage(targets, user, pendingDamage, false, element, adventure);
+	}
+).setTargetingTags({ target: `blast${SAFE_DELIMITER}1`, team: "foe", needsLivingTargets: true })
+	.setUpgrades("Vexing Prismatic Blast")
+	.setDurability(15)
+	.setDamage(40);